Implemented delete user and delete permission capabilities to CDW. Also fixed several UI flow issues.
SVN-Revision: 18697

+++ cacisweb/src/main/java/gov/nih/nci/cacisweb/action/CdwUserDeleteAction.java
@@ -2,6 +2,7 @@ package gov.nih.nci.cacisweb.action;
 
 import gov.nih.nci.cacisweb.dao.DAOFactory;
 import gov.nih.nci.cacisweb.dao.ICDWUserPermissionDAO;
+import gov.nih.nci.cacisweb.exception.DAOException;
 import gov.nih.nci.cacisweb.model.CdwPermissionModel;
 import gov.nih.nci.cacisweb.model.CdwUserModel;
 
@@ -16,16 +17,25 @@ public class CdwUserDeleteAction extends ActionSupport {
     private static final long serialVersionUID = 1L;
 
     private CdwUserModel cdwUserBean;
-    
+
     private CdwPermissionModel cdwPermissionBean;
-    
+
     @Override
     public String execute() throws Exception {
         log.debug("execute() - START");
-        DAOFactory daoFactory = DAOFactory.getDAOFactory();
-        ICDWUserPermissionDAO cdwUserPermissionDAO = daoFactory.getCDWUserPermissionDAO();
-        if(cdwUserPermissionDAO.deleteUser(getCdwUserBean()) > 0){
-            addActionMessage(getText("cdwUserBean.deleteUserSuccessful"));
+        try {
+            DAOFactory daoFactory = DAOFactory.getDAOFactory();
+            ICDWUserPermissionDAO cdwUserPermissionDAO = daoFactory.getCDWUserPermissionDAO();
+            if (!cdwUserPermissionDAO.isUserExists(getCdwUserBean())) {
+                addActionError(getText("cdwUserBean.usernameDoesNotExist"));
+                return INPUT;
+            }
+            if (cdwUserPermissionDAO.deleteUser(getCdwUserBean()) > 0) {
+                addActionMessage(getText("cdwUserBean.deleteUserSuccessful"));
+            } 
+        } catch (DAOException daoe) {
+            addActionError(daoe.getMessage());
+            return ERROR;
         }
         log.debug("execute() - END");
         return INPUT;
@@ -39,15 +49,12 @@ public class CdwUserDeleteAction extends ActionSupport {
         this.cdwUserBean = cdwUserBean;
     }
 
-    
     public CdwPermissionModel getCdwPermissionBean() {
         return cdwPermissionBean;
     }
 
-    
     public void setCdwPermissionBean(CdwPermissionModel cdwPermissionBean) {
         this.cdwPermissionBean = cdwPermissionBean;
     }
-    
 
 }
